edit done all

+++ asr_toolkit/data/datamodule.py
@@ -66,16 +66,16 @@ class DataModule(pl.LightningDataModule):
         Take feature and input, transform and then padding it
         """
         specs = [i[0] for i in batch]
-        input_lengths = torch.LongTensor([i.size(0) for i in specs])
+        input_lengths = torch.IntTensor([i.size(0) for i in specs])
         trans = [i[1] for i in batch]
 
         # batch, time, feature
         specs = torch.nn.utils.rnn.pad_sequence(specs, batch_first=True)
 
         trans = [self.text_process.text2int(s) for s in trans]
-        target_lengths = torch.LongTensor([s.size(0) for s in trans])
+        target_lengths = torch.IntTensor([s.size(0) for s in trans])
         trans = torch.nn.utils.rnn.pad_sequence(trans, batch_first=True).to(
-            dtype=torch.long
+            dtype=torch.int
         )
 
         return specs, input_lengths, trans, target_lengths
